changed methods of saving and retrieving shared preferences all throughout the app
busisness plan attributes for the current user are stored as shared
preferences with a key that reflects their name, e.g. "name": String
name.

Once a user has submitted their plan, we will construct a BusinessPlan
object and save it to the SharedPreferences using the key "Business
Plan". This part still needs to be implemented.